+++ R/ci.R
@@ -115,6 +115,12 @@ ci.mlm <- function(x, ci = .95, ...) {
 
 
 #' @export
+ci.bayesx <- function(x, ci = .95, ...) {
+  ci_wald(model = x, ci = ci, dof = Inf, robust = FALSE, component = "conditional")
+}
+
+
+#' @export
 ci.averaging <- function(x, ci = .95, component = c("conditional", "full"), ...) {
   component <- match.arg(component)
   ci_wald(model = x, ci = ci, dof = Inf, component = component)