+++ R/dof.R
@@ -112,7 +112,7 @@ dof <- degrees_of_freedom
   info <- insight::model_info(model, verbose = FALSE)
 
   ## TODO remove is.list() when insight 0.8.3 on CRAN
-  if (!is.null(info) && is.list(info) && info$is_bayesian) {
+  if (!is.null(info) && is.list(info) && info$is_bayesian && !inherits(model, "bayesx")) {
     model <- bayestestR::bayesian_as_frequentist(model)
   }